requirements (1).txt
absl-py==1.4.0
accelerate==0.22.0
ailment==9.2.57
aiohttp==3.8.5
aiosignal==1.3.1
angr==9.2.57
angr-management==9.2.57
antlr4-python3-runtime==4.8
archinfo==9.2.57
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
bitarray==2.8.1
bitstring==4.0.2
cachetools==5.3.0
capstone==4.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
claripy==9.2.57
cle==9.2.57
click==8.1.7
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
CppHeaderParser==2.7.4
cycler==0.11.0
Cython==3.0.2
datasets==2.14.5
debugpy==1.6.7
decorator==4.4.2
dill==0.3.7
docopt==0.6.2
dpkt==1.9.8
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
executing==1.2.0
fairseq @ git+https://github.com/pytorch/fairseq.git@b5d89cddc9e4a0af831d2aafc1ba7dbf0f1b10d0
fastwer==0.1.3
filelock==3.12.3
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
h5py==3.8.0
html5lib==1.1
huggingface-hub==0.16.4
hydra-core==1.0.7
idna==3.4
imageio==2.28.1
imgaug==0.4.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
itanium-demangler==1.1
jax==0.4.10
jedi==0.18.2
Jinja2==3.1.2
jiwer==3.0.3
joblib==1.1.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
karateclub==1.3.3
keras==2.12.0
Keras-Applications==1.0.8
keras-ocr==0.8.9
kiwisolver==1.4.4
lazy_loader==0.2
Levenshtein==0.21.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
mulpyplexer==0.9
multidict==6.0.4
multiprocess==0.70.15
nampa==0.1.1
natsort==8.4.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
omegaconf==2.0.6
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.5.1
plumbum==1.8.2
ply==3.11
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.23.1
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pyclipper==1.3.0.post4
pycparser==2.21
pyelftools==0.29
Pygments==2.15.1
PyGSP==0.5.1
pyparsing==3.0.9
pyqodeng==0.0.9
PySide6-Essentials==6.5.0
PySide6-QtAds==4.0.3
PySMT==0.9.5
PySocks==1.7.1
pyspellchecker==0.7.2
pytesseract==0.3.10
python-dateutil==2.8.2
python-Levenshtein==0.21.0
python-louvain==0.16
pytz==2023.3
pyvex==9.2.57
PyWavelets==1.4.1
pywin32==306
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.0
QtAwesome==1.2.3
qtconsole==5.4.1
QtPy==2.3.1
qtterm==0.5.1
rapidfuzz==3.0.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rpyc==5.3.1
rsa==4.9
sacrebleu==2.3.1
safetensors==0.3.3
scikit-image==0.20.0
scikit-learn==1.0.2
scipy==1.10.1
sentencepiece==0.1.99
shapely==2.0.1
shiboken6==6.5.0
six==1.16.0
sklearn==0.0.post5
smart-open==6.3.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboardX==2.6.2.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thefuzz==0.19.0
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.7
tokenizers==0.13.3
tomlkit==0.11.8
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.33.1
typing_extensions==4.7.1
tzdata==2023.3
unicorn==2.0.1.post1
urllib3==1.26.15
validators==0.22.0
Wand==0.6.11
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.4
widgetsnbextension==4.0.7
wrapt==1.14.1
xxhash==3.3.0
yarg==0.1.9
yarl==1.9.2
z3-solver==4.10.2.0
